+++ test/02.onComplete.test.js
@@ -37,24 +37,28 @@ describe ( 'askForPromise - "onComplete" function', () => {
                  var
                   list = [ 2, 5, 1, 4 ]
                 , res  = []
+                , i = 0
                 ;
 
             let task = askForPromise ( list )
-
-            list.forEach ( ( el, i ) => {
-                                           setTimeout ( () => {
-                                                                  res.push ( el )
-                                                                  task[i].done ( el )
-                                              }, el )
-                 })
+            task.each ( ({value, done}) => {
+                          setTimeout ( () => { 
+                                            res.push ( value )
+                                            done ( i )
+                                            i ++
+                                }, value )
+                    })
 
             task
              .onComplete ( r => {
                                         expect ( res ).to.have.length ( 4 )
                                         expect ( r   ).to.have.length ( 4 )
-                                        
+                            
                                         expect ( r[0]).to.equal (list[0] )
                                         expect ( res[0]).to.not.equal (list[0] )
+
+                                        expect ( res ).to.be.deep.equal ( list )
+                                        expect ( r ).to.be.deep.equal ( [1,3,0,2] )
                  })
        }) // it done
 
@@ -75,13 +79,13 @@ describe ( 'askForPromise - "onComplete" function', () => {
 
     it ( 'onComplete: Alternative promise race', () => {
           var
-                  list = [ 2, 5, 1, 4 ]
+                  list = [ 20, 50, 1, 40 ]
                 , res  = []
                 ;
 
             let task = askForPromise ()
 
-            list.forEach ( ( el, i ) => setTimeout ( () => {
+            list.forEach ( el  => setTimeout ( () => {
                                                               res.push(el)
                                                               task.done ( el )
                                               } , el )