+++ types/askForPromise.d.ts
@@ -1,12 +1,12 @@
 export default askForPromise;
-export type AskObject = any;
 /**
- * @typedef {Object} AskObject
+ * @interface AskObject
  * @description Object with promise and related helper functions
  * @property {Promise} [promise] - Promise object if a single promise is created
- * @property {Array<Promise>} [promises] - Array of promises if multiple promises are created
+ * @property {Array<AskObject>|null} [promises] - Array of promises if multiple promises are created
  * @property {Function} done - Resolve function
  * @property {Function} cancel - Reject function
+ * @property {Function} each - Callback function to be called for each list item
  * @property {Function} onComplete - Function to be called after promise is resolved
  * @property {Function} timeout - Function to set timeout on promise
  */
@@ -23,16 +23,16 @@ declare namespace askForPromise {
      * @description Executes list of functions that return a promise in sequence.
      * @param {Array<Function>} list - List of functions that return a promise
      * @param {...any} args - Arguments to be passed to each function in the list
-     * @returns {askObject} - Object with promise and related helper functions
+     * @returns {AskObject} - Object with promise and related helper functions
      */
-    function sequence(list: Array<Function>, ...args: any[]): askObject;
+    function sequence(list: Array<Function>, ...args: any[]): AskObject;
     /**
      * @function all
      * @description Executes list of functions that return a promise in parallel.
      * @param {Array<Function>} list - List of functions that return a promise
      * @param {...any} args - Arguments to be passed to each function in the list
-     * @returns {askObject} - Object with promise and related helper functions
+     * @returns {AskObject} - Object with promise and related helper functions
      */
-    function all(list: Array<Function>, ...args: any[]): askObject;
+    function all(list: Array<Function>, ...args: any[]): AskObject;
 }
 //# sourceMappingURL=askForPromise.d.ts.map